pipeline-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
requests==2.32.3
pydantic==2.7.4
